refactor(backend): split out oauth session routes into a submodule
